src/aibc/memory/cognee-service/requirements.txt
cognee[gemini]
fastapi
uvicorn[standard]
python-dotenv
